requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Data
albumentations>=1.3.0
opencv-python>=4.7.0
Pillow>=9.5.0

# Training
pyyaml>=6.0
tensorboard>=2.13.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Export
onnx>=1.14.0
onnxruntime>=1.15.0

# TensorRT (install separately on Jetson)
# tensorrt>=8.6.0
# pycuda>=2022.1

# ROS2 (install via apt on target system)
# rclpy
# sensor_msgs
# cv_bridge

# Optional
# wandb>=0.15.0
scikit-learn>=1.2.0
